% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{YoloGrandJury}
\alias{YoloGrandJury}
\title{Yolo County grand jury data}
\format{A list with three data frames representing estimated eligible
population, grand jury nominees, and grand jurors.}
\usage{
data(YoloGrandJury)
}
\description{
A dataset containing describing the composition of the Yolo County
grand jury over the years 1985--2004.  This dataset comes
as a list of three data frames:
}
\details{
\emph{Eligible}:  This data frame contains estimates of the
total number of residents of Yolo County who are eligible to serve
on the grand jury and of the number of residents who were Hispanic.

\emph{Nominee}:  This data frame contains the number of grand jury nominees
selected from the grand jury applicant pool as well as the estimated number
of grand jury nominees who were Hispanic.

\emph{Juror}:  This data frame contains the number of grand jurors
seated from the grand jury nominee pool as well as the estimated number
of grand jurors who were Hispanic.

The source of data and the methods of estimation are described in:

Braun, J.V. (2007) Exhibit C:  Statistical and probabilistic analysis
of representation of Hispanics and Asian & Pacific Islanders in the
Yolo County grand jury from 1985 to 2004.
\emph{The People of the State of California vs. Michael Raquel, et al.,
Case No. 01-1577, Superior Court of the State of California
in and for the County of Yolo.}
}
\keyword{datasets}

